# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplyTemplateIn": {
        "properties": {
          "layout_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout Id"
          },
          "style_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Id"
          }
        },
        "title": "ApplyTemplateIn",
        "type": "object"
      },
...
      "ArtworkPageOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArtworkSummaryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ArtworkPageOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SnapshotIn": {
        "properties": {
          "reason": {
            "default": "manual",
            "enum": [
              "opened",
              "manual"
            ],
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "SnapshotIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "the_frame_v2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/artworks": {
      "get": {
        "description": "The simple cases as query parameters; a full filter AST goes to `POST /artworks/query`.",
        "operationId": "list_artworks_api_v1_artworks_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "draft",
                    "ready"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "favorite",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Favorite"
            }
          },
          {
            "in": "query",
            "name": "photo_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Photo Id"
            }
          },
          {
            "in": "query",
            "name": "collection_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Collection Id"
            }
          },
          {
            "in": "query",
            "name": "include_nested",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Nested",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "created_desc",
              "enum": [
                "created_desc",
                "created_asc",
                "updated_desc",
                "title_asc",
                "manual"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Artworks",
        "tags": [
          "artworks"
        ]
      },
      "post": {
        "operationId": "create_artwork_api_v1_artworks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtworkCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Artwork",
        "tags": [
          "artworks"
        ]
      }
    },
    "/api/v1/artworks/query": {
      "post": {
        "description": "The filter bar and smart collections: one AST (docs/data-model.md \u00a75.2), one page back.",
        "operationId": "query_artworks_api_v1_artworks_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtworkQueryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Artworks",
        "tags": [
          "artworks"
        ]
      }
    },
    "/api/v1/artworks/{artwork_id}": {
      "delete": {
        "description": "Move to trash (restore and purge: `/trash`, docs/organization.md \u00a75).",
        "operationId": "trash_artwork_api_v1_artworks__artwork_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trash Artwork",
        "tags": [
          "artworks"
        ]
      },
      "get": {
        "operationId": "get_artwork_api_v1_artworks__artwork_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artwork",
        "tags": [
          "artworks"
        ]
      },
      "patch": {
        "operationId": "update_artwork_api_v1_artworks__artwork_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtworkUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Artwork",
        "tags": [
          "artworks"
        ]
      }
    },
    "/api/v1/artworks/{artwork_id}/apply-template": {
      "post": {
        "description": "Re-dress and/or re-lay out an artwork from a saved template (docs/templates.md \u00a75).\n\nCopy on apply: the template is read once and the artwork records it as its origin. A\n`pre_template_update` snapshot is taken first, so the change is undoable from the history.",
        "operationId": "apply_template_api_v1_artworks__artwork_id__apply_template_post",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyTemplateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Template",
        "tags": [
          "artworks"
        ]
      }
    },
    "/api/v1/artworks/{artwork_id}/document": {
      "put": {
        "description": "Replace the document. `If-Match: <document_version>` is required (409 when outdated).",
        "operationId": "put_document_api_v1_artworks__artwork_id__document_put",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtworkDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Document",
        "tags": [
          "artworks"
        ]
      }
    },
    "/api/v1/artworks/{artwork_id}/duplicate": {
      "post": {
        "operationId": "duplicate_artwork_api_v1_artworks__artwork_id__duplicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Duplicate Artwork",
        "tags": [
          "artworks"
        ]
      }
    },
    "/api/v1/artworks/{artwork_id}/render.jpg": {
      "get": {
        "operationId": "render_jpg_api_v1_artworks__artwork_id__render_jpg_get",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          },
          {
            "description": "Render hash: the response is cached forever when it matches",
            "in": "query",
            "name": "v",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Render hash: the response is cached forever when it matches",
              "title": "V"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Render Jpg",
        "tags": [
          "artworks"
        ]
      }
    },
    "/api/v1/artworks/{artwork_id}/render.png": {
      "get": {
        "description": "Lossless 3840\u00d72160 master (rendered on demand when missing).",
        "operationId": "render_png_api_v1_artworks__artwork_id__render_png_get",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          },
          {
            "description": "Render hash: the response is cached forever when it matches",
            "in": "query",
            "name": "v",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Render hash: the response is cached forever when it matches",
              "title": "V"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Render Png",
        "tags": [
          "artworks"
        ]
      }
    },
    "/api/v1/artworks/{artwork_id}/snapshots": {
      "get": {
        "operationId": "list_snapshots_api_v1_artworks__artwork_id__snapshots_get",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SnapshotOut"
                  },
                  "title": "Response List Snapshots Api V1 Artworks  Artwork Id  Snapshots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Snapshots",
        "tags": [
          "artworks"
        ]
      },
      "post": {
        "operationId": "create_snapshot_api_v1_artworks__artwork_id__snapshots_post",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Snapshot",
        "tags": [
          "artworks"
        ]
      }
    },
    "/api/v1/artworks/{artwork_id}/snapshots/{snapshot_id}/restore": {
      "post": {
        "operationId": "restore_snapshot_api_v1_artworks__artwork_id__snapshots__snapshot_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Snapshot",
        "tags": [
          "artworks"
        ]
      }
    },
...
    "/api/v1/artworks/{artwork_id}/validate": {
      "post": {
        "description": "Mark as ready (displayable). 422 `artwork_incomplete` / `invalid_document` otherwise.",
        "operationId": "validate_artwork_api_v1_artworks__artwork_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "artwork_id",
            "required": true,
            "schema": {
              "title": "Artwork Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Artwork",
        "tags": [
          "artworks"
        ]
      }
    },
...
  }
}